TdpAgent/requirements.txt
# sqlalchemy













# annotated-doc==0.0.4
# annotated-types==0.7.0
# anyio==4.12.0
# certifi==2025.11.12
# cffi==2.0.0
# charset-normalizer==3.4.4
# click==8.3.1
# colorama==0.4.6
# cryptography==46.0.3
# distro==1.9.0
# dnspython==2.8.0
# email-validator==2.3.0
# fastapi==0.124.4
# greenlet==3.3.0
# h11==0.16.0
# httpcore==1.0.9
# httpx==0.28.1
# idna==3.11
# jiter==0.12.0
# lxml==6.0.2
# openai==2.11.0
# pdfminer.six==20251107
# pycparser==2.23
# pydantic==2.12.5
# pydantic_core==2.41.5
# PyMySQL==1.1.2
# python-docx==1.2.0
# python-dotenv==1.2.1
# requests==2.32.5
# sniffio==1.3.1
# SQLAlchemy==2.0.45
# starlette==0.50.0
# tqdm==4.67.1
# typing-inspection==0.4.2
# typing_extensions==4.15.0
# urllib3==2.6.2
# uvicorn==0.38.0


























# annotated-doc==0.0.4
# annotated-types==0.7.0
# anyio==4.12.0
# certifi==2025.11.12
# cffi==2.0.0
# charset-normalizer==3.4.4
# click==8.3.1
# colorama==0.4.6
# cryptography==46.0.3
# distro==1.9.0
# dnspython==2.8.0
# email-validator==2.3.0
# fastapi==0.124.4
# greenlet==3.3.0
# h11==0.16.0
# httpcore==1.0.9
# httpx==0.28.1
# idna==3.11
# jiter==0.12.0
# lxml==6.0.2
# openai==2.11.0
# pdfminer.six==20251107
# pycparser==2.23
# pydantic==2.12.5
# pydantic_core==2.41.5
# PyMySQL==1.1.2
# python-docx==1.2.0
# python-dotenv==1.2.1
# requests==2.32.5
# sniffio==1.3.1
# SQLAlchemy==2.0.45
# starlette==0.50.0
# tqdm==4.67.1
# typing-inspection==0.4.2
# typing_extensions==4.15.0
# urllib3==2.6.2
# uvicorn==0.38.0
# pyodbc==4.0.35




























annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.124.4
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
lxml==6.0.2
openai==2.11.0
pdfminer.six==20251107
pillow==12.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.7
PyMySQL==1.1.2
python-docx==1.2.0
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
